+++ frontend/src/views/CategoryListPage.vue
@@ -3,7 +3,7 @@
     <br />
 
     <div>
-      <router-link class="no-decoration" v-for="category in categories" :key="category.key" :to="{ name: 'category-detail', params: { categoryKey: category.key } }">
+      <router-link class="no-decoration" v-for="category in categories" :key="category.name" :to="{ name: 'category-detail', params: { categoryName: category.name } }">
         <span class="category">
           <h3>{{ category.name }}</h3>
           <p><strong>{{ category.question_count }}</strong> question<span v-if="category.question_count > 1">s</span></p>
@@ -13,8 +13,20 @@
 
     <br />
     <hr v-if="categories" />
-    <div v-if="categories" class="row actions justify-content-end">
-      <div class="col-sm-4">
+    <div v-if="categories" class="row actions">
+      <div class="col-sm">
+        <router-link :to="{ name: 'question-list' }">
+          Toutes les questions
+        </router-link>
+        <br />
+      </div>
+      <div class="col-sm">
+        <router-link :to="{ name: 'about' }">
+          ℹ️&nbsp;À propos de cette application
+        </router-link>
+        <br />
+      </div>
+      <div class="col-sm">
         <HomeLink />
       </div>
     </div>