+++ frontend/src/views/QuestionDetailPage.vue
@@ -88,7 +88,7 @@
         </router-link>
       </div>
       <div class="col-sm">
-        <router-link :to="{ name: 'category-detail', params: { categoryKey: question.category }  }">
+        <router-link :to="{ name: 'category-detail', params: { categoryName: question.category }  }">
           🏷️&nbsp;Toutes les questions <span class="text-secondary">{{ question.category }}</span>
         </router-link>
       </div>